Remote tld dependency, rename canvas_exceptions to canvas_allowed_hosts.
Dependency for tld was only needed to get TLD+1 for URLs exempt
from canvas defense. Yet, we don't need the canvas_exceptions for
TBB >= 4.5a3 and we can expect the caller to pass allowed hosts
instead of URLs if they need to use TorBrowserDriver with outdated TB
versions.

We rename the argument canvas_exceptions to canvas_allowed_hosts to
signal the change.